Let's call API call response `movie_dict`
I think the line that we should try to remove is `d=json.dumps(movie_dict,indent=-10)`.

The API call has returned a nicely formatted dict which is just about the coolest data structure in Python.  Instead of deconstructing the dict into a big long string and then searching that string, we should just stick with the dict.  My sense is that `queryData()` should return a dict and that other functions like `mineData()` should be rewritten to accept a dict instead of text.